Hack around cargo limitations with a build.rs
See comments in the code or
https://github.com/jgallagher/rusqlite/pull/608 for more griping.
