+++ components/remerge/Cargo.toml
@@ -23,6 +23,14 @@ sync-guid = {path = "../support/guid", features = ["random", "rusqlite_support"]
 matches = "0.1.8"
 num-traits = "0.2.8"
 
+# A *direct* dep on the -sys crate is required for our build.rs
+# to see the DEP_SQLITE3_LINK_TARGET env var that cargo sets
+# on its behalf.
+libsqlite3-sys = "0.17.0"
+
 [dependencies.rusqlite]
 version = "0.21.0"
 features = ["functions", "bundled", "serde_json"]
+
+[build-dependencies]
+nss_build_common = { path = "../support/rc_crypto/nss/nss_build_common" }
