+++ components/support/rc_crypto/nss/nss_sys/Cargo.toml
@@ -13,3 +13,4 @@ bindgen = "0.52.0"
 serde = "1.0.101"
 serde_derive = "1.0.101"
 toml = "0.5.3"
+nss_build_common = {path = "../nss_build_common"}
