+++ components/support/rc_crypto/nss/nss_sys/build.rs
@@ -4,21 +4,12 @@
 
 use bindgen::Builder;
 use serde_derive::Deserialize;
-use std::{
-    env,
-    ffi::OsString,
-    fs,
-    path::{Path, PathBuf},
-    process::Command,
-};
+use std::{env, fs, path::PathBuf, process::Command};
 use toml;
 
-const BINDINGS_CONFIG: &'static str = "bindings.toml";
+use nss_build_common::*;
 
-enum LinkingKind {
-    Dynamic { folded_libs: bool },
-    Static,
-}
+const BINDINGS_CONFIG: &'static str = "bindings.toml";
 
 // This is the format of a single section of the configuration file.
 #[derive(Deserialize)]
@@ -46,30 +37,6 @@ struct Bindings {
     exclude: Option<Vec<String>>,
 }
 
-fn env(name: &str) -> Option<OsString> {
-    println!("cargo:rerun-if-env-changed={}", name);
-    env::var_os(name)
-}
-
-fn env_str(name: &str) -> Option<String> {
-    println!("cargo:rerun-if-env-changed={}", name);
-    env::var(name).ok()
-}
-
-fn env_flag(name: &str) -> bool {
-    match env_str(name).as_ref().map(String::as_ref) {
-        Some("1") => true,
-        Some("0") => false,
-        Some(s) => {
-            println!(
-                "cargo:warning=unknown value for environment var {:?}: {:?}. Ignoring",
-                name, s
-            );
-            false
-        }
-        None => false,
-    }
-}
 const DEFAULT_ANDROID_NDK_API_VERSION: &str = "21";
 
 // Set the CLANG_PATH env variable to point to the right clang for the NDK in question.
@@ -103,100 +70,13 @@ fn main() {
     // Note: this has to be first!
     maybe_setup_ndk_clang_path();
     // 1. NSS linking.
-    let (lib_dir, include_dir) = get_nss();
-    println!(
-        "cargo:rustc-link-search=native={}",
-        lib_dir.to_string_lossy()
-    );
-    println!("cargo:include={}", include_dir.to_string_lossy());
-    let kind = determine_kind();
-    link_nss_libs(&kind);
+    let (_, include_dir) = link_nss().expect("To build nss_sys, NSS_DIR must be set!");
     // 2. Bindings.
     let config_file = PathBuf::from(BINDINGS_CONFIG);
     println!("cargo:rerun-if-changed={}", config_file.to_str().unwrap());
     let config = fs::read_to_string(config_file).expect("unable to read binding configuration");
     let bindings: Bindings = toml::from_str(&config).unwrap();
     build_bindings(&bindings, &include_dir.join("nss"));
-}
-
-fn determine_kind() -> LinkingKind {
-    return if env_flag("NSS_STATIC") {
-        LinkingKind::Static
-    } else {
-        let folded_libs = env_flag("NSS_USE_FOLDED_LIBS");
-        return LinkingKind::Dynamic { folded_libs };
-    };
-}
-
-fn link_nss_libs(kind: &LinkingKind) {
-    let libs = get_nss_libs(kind);
-    // Emit -L flags
-    let kind_str = match kind {
-        LinkingKind::Dynamic { .. } => "dylib",
-        LinkingKind::Static => "static",
-    };
-    for lib in libs {
-        println!("cargo:rustc-link-lib={}={}", kind_str, lib);
-    }
-}
-
-fn get_nss_libs(kind: &LinkingKind) -> Vec<&'static str> {
-    match kind {
-        LinkingKind::Static => {
-            let mut static_libs = vec![
-                "certdb",
-                "certhi",
-                "cryptohi",
-                "freebl_static",
-                "hw-acc-crypto",
-                "nspr4",
-                "nss_static",
-                "nssb",
-                "nssdev",
-                "nsspki",
-                "nssutil",
-                "pk11wrap_static",
-                "plc4",
-                "plds4",
-                "softokn_static",
-            ];
-            // Hardware specific libs.
-            let target_arch = env::var("CARGO_CFG_TARGET_ARCH").unwrap();
-            let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap();
-            // https://searchfox.org/mozilla-central/rev/1eb05019f47069172ba81a6c108a584a409a24ea/security/nss/lib/freebl/freebl.gyp#159-168
-            if target_arch == "x86_64" || target_arch == "x86" {
-                static_libs.push("gcm-aes-x86_c_lib");
-            } else if target_arch == "aarch64" {
-                static_libs.push("gcm-aes-aarch64_c_lib");
-            }
-            // https://searchfox.org/mozilla-central/rev/1eb05019f47069172ba81a6c108a584a409a24ea/security/nss/lib/freebl/freebl.gyp#224-233
-            if ((target_os == "android" || target_os == "linux") && target_arch == "x86_64")
-                || target_os == "windows"
-            {
-                static_libs.push("intel-gcm-wrap_c_lib");
-                // https://searchfox.org/mozilla-central/rev/1eb05019f47069172ba81a6c108a584a409a24ea/security/nss/lib/freebl/freebl.gyp#43-47
-                if (target_os == "android" || target_os == "linux") && target_arch == "x86_64" {
-                    static_libs.push("intel-gcm-s_lib");
-                }
-            }
-            return static_libs;
-        }
-        LinkingKind::Dynamic { folded_libs } => {
-            let mut dylibs = vec!["freebl3", "nss3", "nssckbi", "softokn3"];
-            if !folded_libs {
-                dylibs.append(&mut vec!["nspr4", "nssutil3", "plc4", "plds4"]);
-            }
-            return dylibs;
-        }
-    }
-}
-
-fn get_nss() -> (PathBuf, PathBuf) {
-    let nss_dir = env("NSS_DIR").expect("To build nss_sys, NSS_DIR must be set!");
-    let nss_dir = Path::new(&nss_dir);
-    let lib_dir = nss_dir.join("lib");
-    let include_dir = nss_dir.join("include");
-    (lib_dir, include_dir)
 }
 
 fn build_bindings(bindings: &Bindings, include_dir: &PathBuf) {
